walker_app_api/pyproject.toml
[project]
name = "walker-app-api"
version = "0.1.0"
description = "FastAPI backend for AI-Now content aggregation"
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pydantic[email]>=2.10",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.9",
    "httpx>=0.27",
    "requests>=2.32.2",
    "sqlalchemy==2.0.23",
    "alembic==1.13.1",
    "psycopg2-binary==2.9.9",
    "feedparser==6.0.10",
    "beautifulsoup4==4.12.2",
    "lxml>=5.3.0",
    "numpy>=1.25.0",
    "pydantic-ai==1.0.15",
    "tavily-python>=0.7.12",
    "scikit-learn>=1.3.0",
    "python-jose[cryptography]==3.3.0",
    "passlib[bcrypt]==1.7.4",
    "nltk>=3.8",
    "textstat>=0.7.0",
    "sentence-transformers>=2.2.0",
    "cloudscraper==1.2.71",
    "python-dateutil>=2.8.2",
    "selenium>=4.10.0",
    "webdriver-manager>=4.0.1",
    "gunicorn>=21.2.0",
    "pytest>=7.4.0",
    "argon2-cffi>=23.1.0",
    "PyJWT>=2.9.0",
]

[tool.uv]
dev-dependencies = []